extract group session to separate component
to support admin preview
